Refactoring of all outfits pages
