+++ app/(tabs)/outfits.tsx
@@ -1,16 +1,9 @@
-import { Button, ScrollView, StyleSheet, Platform, Dimensions} from 'react-native';
-import { Text, View } from '@/components/Themed';
-
-import React, {useState} from 'react';
-import {Alert, Modal, Pressable, FlatList, TouchableOpacity } from 'react-native';
-import {SafeAreaView, TextInput} from 'react-native';
-
-import { Clothe } from '@/classes/clothe';
-import {useNavigation} from "@react-navigation/native";
-import {Picker} from '@react-native-picker/picker';
-
-import EditOutfits from '@/app/edit'
-import { BottomSheet } from 'react-native-elements';
+import { ScrollView, StyleSheet, Platform } from 'react-native';
+import { View } from '@/components/Themed';
+import React from 'react';
+import {searchPlaceHolder} from '@/components/SearchBarPlaceholder'
+import {addOutfitButton} from '@/components/OutfitButton'
+import {outfitFilterBoxes} from '@/components/OutfitsFilterBoxes';
 
 /* To Do: 
 
@@ -72,79 +65,17 @@ GitHub Issues:
   belongs to? Ej. Type would be [Type] instead of 
                                                       Type
                                                     [Unknown]
+
+  Done: Refactoring of all outfits pages
   
 */
 
-
-// Add Outfit Button
-function addOutfitButton(){
-
-  return(
-    <View style={{paddingBottom:5}}>
-    {/* Separator */}
-      <View style={styles.separator} lightColor="#ffff" darkColor="rgba(255,255,255,0.1)" />
-
-    {/* Button */}
-      <View style={styles.OutfitButton}> 
-        <Pressable>
-            <Text style={styles.OutfitTButtonTextStyle}>Outfit</Text>
-        </Pressable>
-      </View>
-
-
-    </View>
-
-  );
-}
-
-function searchPlaceHolder(){
-  const [number, onChangeNumber] = React.useState('');
-
-  return(
-    <SafeAreaView>
-      <TextInput
-      style={{
-        paddingHorizontal: 5,
-        color:"black", 
-        // fontWeight:'bold'
-      }}
-        onChangeText={onChangeNumber}
-        value={number}
-        placeholder="Search"
-        // keyboardType="numeric"
-        placeholderTextColor="#9f9f9f"
-      />
-    </SafeAreaView>
-  );
-}
-
-
 export default function TabTwoScreen() {
-
-  const [selectedType, setSelectedType] = useState(Clothe.Type.Unknown);
-  const [selectedColor, setSelectedColor] = useState(Clothe.Color.Unknown); // Default value
-  const [selectedSize, setSelectedSize] = useState(Clothe.SleeveSize.Unknown); // Default value
-  
-  const typeKeys = Object.keys(Clothe.Type).filter(key => isNaN(Number(key)));
-  const colorKeys = Object.keys(Clothe.Color).filter(key => isNaN(Number(key)));
-  const sleeveSizeKeys = Object.keys(Clothe.SleeveSize).filter(key => isNaN(Number(key)));
-
-
-  const navigation = useNavigation();
-
-  const handleCancel = () => {
-    // Navigate back to the previous screen or any desired screen
-    navigation.goBack();
-  };
-
-  const screenWidth = Dimensions.get('window').width;
-  const pickerWidth = screenWidth / 3; // Assuming you have 3 Pickers
 
   return (
 
     <View style={styles.bigContainer}>
     {/* Bigger Container */}
-
 
     <View style={styles.fixedContainer}> 
       {/* Search & Filter container is inside this Fixed container */}
@@ -154,150 +85,40 @@ export default function TabTwoScreen() {
 
         {/* Search Bar */}
         <View style={styles.searchContainer}>
-          {/* <Text style={styles.searchText}>Search</Text> */}
           {searchPlaceHolder()}
-
+          
         </View>
-            <View style={styles.filterTextContainer}>              
-              <Text style={styles.filterText}>Type</Text>
-              <Text style={styles.filterText}>Color</Text>
-              <Text style={styles.filterText}>Sleeve Size</Text>
 
-            </View>
-        {/* Filter boxes */}
-        <View style={styles.pickerBox}>
+        {/* Filter Titles and Boxes */}
+        {outfitFilterBoxes()}
 
-            {/* <View style={styles.filterContainer}>   */}
-
-                  <Picker style={styles.picker} itemStyle={styles.pick}
-                      selectedValue={selectedType} 
-                      onValueChange={(itemValue: Clothe.Type) => 
-                          setSelectedType(itemValue) // Update the selectedType state 
-                  }>
-                  {typeKeys.map((typeKey) => (
-                      <Picker.Item
-                          key={typeKey}
-                          label={typeKey}
-                          value={Clothe.Type[typeKey as keyof typeof Clothe.Type]}
-                      />
-                      ))}
-                  </Picker>
-            {/* </View> */}
-
-
-
-            {/* <View style={styles.filterContainer}>   */}
-                {/* <Text style={styles.filterText}>Color</Text> */}
-                    <Picker style={styles.picker} itemStyle={styles.pick}
-                        selectedValue={selectedColor} 
-                        onValueChange={(itemValue) => 
-                            setSelectedColor(itemValue)
-                        }>
-                        {colorKeys.map((colorKey) => (
-                            <Picker.Item
-                                key={colorKey}
-                                label={colorKey}
-                                value={Clothe.Color[colorKey as keyof typeof Clothe.Color]}
-                            />
-                        ))}
-                    </Picker>
-            {/* </View> */}
-
-
-
-            {/* <View style={styles.filterContainer}>   */}
-                {/* <Text style={styles.filterText}>Sleeve Size</Text> */}
-                    <Picker style={styles.picker} itemStyle={styles.pick}
-                        selectedValue={selectedSize}
-                        onValueChange={(itemValue, itemIndex) => 
-                            setSelectedSize(itemValue)
-                        }>
-                        {sleeveSizeKeys.map((sleeveSizeKey) => (
-                            <Picker.Item
-                                key={sleeveSizeKey}
-                                label={sleeveSizeKey}
-                                value={Clothe.SleeveSize[sleeveSizeKey as keyof typeof Clothe.SleeveSize]}
-                            />
-                        ))}
-                    </Picker>
-              {/* </View> */}
-
-            </View> 
-            {/* filterbox end*/}
-        
         </View> 
         {/* searchAndFilter end*/}
-
-
 
       </View> 
       {/* Fixed Container view */}
 
 
-
-
-
+    {/* All buttons components are inside this smaller container */}
+    {/* Makes it so scrollable content is UNDER fixed container (Filter/SearchBox) */}
     <View style={styles.smallerContainer}> 
-    {/* All components are inside this smaller container */}
-
-
-
-
 
       <ScrollView>
-
-      {/* Outfit Buttons */}
-      <View style={styles.outfitsButtonContainer}> 
+      {/* Outfit Buttons */}      
       {/* All Outfit Button are inside this container */}
-
-
-      {/* First Button without separator */}
-      <View style={{paddingBottom:5}}>
-
-        {/* Button */}
-        <View style={styles.OutfitButton}> 
-            <Pressable  
-            // onPress={() => setModalVisible(true)}
-            >
-                <Text style={styles.OutfitTButtonTextStyle}>OutfitButton1</Text>
-            </Pressable>
-            
-        </View>
-        </View>
-
-
-        {/* All other buttons */}
-
-        {/* Button */}
+      <View style={styles.outfitsButtonContainer}> 
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-        {/* Button */}
         {addOutfitButton()}
-
-
       </View>
-
       </ScrollView>
-
 
     </View> 
     {/* smaller container */}
-
 
     </View> 
     //{/* bigger container */}
@@ -305,36 +126,32 @@ export default function TabTwoScreen() {
   );
 }
 
-
-
 // Styles
 const styles = StyleSheet.create({
   bigContainer: { //all components are in this parent container
     flex: 1,
     // backgroundColor: 'black', //ffff
-
   },
-    
+
   smallerContainer: { //all buttons are inside this smaller container
     flex: 1,
     // backgroundColor: 'pink',
     padding: 10, 
     margin: 5, //margin of the square - how big is it
     borderRadius: 20, //rounds edges
-    top:100,
-    marginBottom:100,
+    top:120,
+    marginBottom:140,
     zIndex: 1, // Ensure it appears above other content
     ...Platform.select({
-      ios: {
-          width: "95%",
-          top:120
-      },
+      // ios: {
+      //     width: "95%",
+      //     top:150
+      // },
       android: {
           width: "95%",
-          top:120,
+          top:150,
       },
   }),
-
   },
 
   fixedContainer: {
@@ -344,32 +161,17 @@ const styles = StyleSheet.create({
     right: 0,
     // bottom:10, //Nope, covers the whole page
     // backgroundColor: 'lightblue',
-    // zIndex: 1, // Ensure it appears above other content
+
     ...Platform.select({
-      ios: {
-          height: "20%",
-          width: "95%",
-      },
-      android: {
-          height: "25%",
-          width: "100%",
-          // top:5
-      },
-  }),
-  },
-
-  searchContainer:{  //search bar is inside this container
-    // flex: 1,
-    backgroundColor: '#D9D9D9',
-    padding: 5, 
-    margin: 10, //margin of the square - how big is it
-    borderRadius: 10, //rounds edges
-    // marginTop: 1,
-    marginVertical: 1,
-  },
-
-  searchText:{  //styling for "Search" text
-  fontWeight:'bold', color:'#9f9f9f', backgroundColor:'#D9D9D9'
+        // ios: {
+        //     height: "33%",
+        //     width: "95%",
+        // },
+        android: {
+            height: "33%",
+            width: "100%",
+        },
+      }),
   },
 
   searchAndFilter:{  //search and filters are inside this container
@@ -403,153 +205,34 @@ const styles = StyleSheet.create({
     }),
 },
 
-filterContainer:{  //filterbox and text label are inside this container
-  // alignItems: 'center',
-  flexDirection: 'row',
-  backgroundColor: '#F0F0F0',
-  flex:1,
-  // paddingHorizontal:12
-  
-},
+  searchContainer:{  //search bar is inside this container
+    backgroundColor: '#D9D9D9',
+    padding: 5, 
+    margin: 10, //margin of the square - how big is it
+    borderRadius: 10, //rounds edges
+    marginVertical: 1,
+  },
 
-filterTextContainer:{
-  flexDirection:'row', 
-  justifyContent:'space-evenly', 
-  backgroundColor: "#F0F0F0", 
-  left:5,
-  ...Platform.select({
-    ios: {
-        width: "33%",
-        bottom: 55,
-    },
-    android: {
-        width: "100%",
-        height: "25%",
-        top: 7,
-        left: 0,
-    },
-  }),
-
-},
-
-filterText: { //styling for *text* on top of filter dropdowns
-  // paddingTop: 10,
-  paddingHorizontal: 10,
-  color:"black", 
-  fontWeight:'bold'
-},
-
-pickerBox: { //container where pickers are located
-    flexDirection: "row",
-    justifyContent: 'space-evenly',
-    // flex:1,
-    height: "45%",
-    backgroundColor:  'rgba(52, 52, 52, 0)', //'lightblue',
-    ...Platform.select({
-      ios: {
-          width: "33%",
-          bottom: 55,
-      },
-      android: {
-          width: "100%",
-          height: "55%",
-          bottom: 0,
-      },
-  }), 
-    
-
-
-
-},
-picker: { // picker outside design
-    borderColor: 'black',
-    borderRadius: 10,
-    // width: pickerWidth, 
-
-    ...Platform.select({
-        ios: {
-            width: "33%",
-            bottom: 55,
-        },
-        android: {
-            width: "40%",
-            height: "55%",
-            bottom: 0,
-        },
-    }),
-},
-pick: { 
-    fontSize: 15,
-    ...Platform.select({
-        ios: {fontSize: 15},
-        android: {fontSize: 0,},
-    }),
-    
-},
-
-
-
-
-
-
-
-
-
-
-
-
-
-  // Outfit Buttons component styles
+  // Outfit Buttons container
   outfitsButtonContainer:{ // all outfits buttons are inside this container
     flex: 1,
     // backgroundColor: 'lightblue',
     padding: 10, 
     margin: 5, //margin of the square - how big is it
     borderRadius: 20, //rounds edges
-    // marginBottom:120,
-    // top:100,
-    // zIndex: 1, // Ensure it appears above other content
-    // maxHeight:480,
-    // paddingBottom: 100,
+    marginTop: -20,
+    
     ...Platform.select({
       ios: {
           height: "50%",
           width: "95%",
       },
       android: {
-          height: "70%",
-          width: "95%",
+          height: "80%",
+          width: "99%",
           
       },
   }),
-
   },
 
-  separator: { //added this to separate each button a smidge, is added just before a new button
-    marginVertical: 10,
-    height: .5,
-    width: '80%',
-    backgroundColor: 'white'
-  },
-
-  OutfitButton: {
-    borderRadius: 15,
-    padding: 20,
-    elevation: 2,
-    shadowColor: '#000',
-    shadowOffset: {
-      width: 0.9,
-      height: 2,
-    },
-    shadowOpacity: 0.25,
-    shadowRadius: 4,    
-    backgroundColor: '#F0F0F0' //'#F194FF' // '#C100E0' //'#f8f4f4'
-  },
-  OutfitTButtonTextStyle: {
-    color: '#C100E0',
-    fontWeight: 'bold',
-    // textAlign: 'center',
-  },
-
-  
 });
