+++ app/edit.tsx
@@ -1,21 +1,13 @@
-import {Button, Pressable, Text, StyleSheet, Alert, TextInput, ScrollView, Platform, KeyboardAvoidingView} from 'react-native';
+import React, {useLayoutEffect} from 'react';
+import {StyleSheet, ScrollView, Platform} from 'react-native';
 import {View} from '@/components/Themed';
 import {useNavigation} from "@react-navigation/native";
-import { MaterialIcons } from '@expo/vector-icons'; // Import icons from Expo vector icons
 
-import {useState} from "react";
-import React from 'react';
-import Icon from 'react-native-vector-icons/MaterialCommunityIcons';
-// import Draggable from '@react-navigation/native';
-// import { PanResponder } from 'react-native';
-// import SortableList from '@react-navigation/native';
-
-import { Clothe } from '@/classes/clothe';
-import {Picker} from '@react-native-picker/picker';
-import {SafeAreaView} from 'react-native';
-
-
-
+import {searchPlaceHolder} from '@/components/SearchBarPlaceholder'
+import {addEditOutfitButton} from '@/components/EditOutfitButton'
+import {CancelButton} from '@/components/CancelButton'
+import {SaveButton} from '@/components/SaveButton'
+import {outfitFilterBoxes} from '@/components/OutfitsFilterBoxes';
 
 /*
     Edit Page 
@@ -27,96 +19,12 @@ import {SafeAreaView} from 'react-native';
     - mover esta in progress
 */
 
-    
 
-// Add Outfit Button
-function addOutfitButton(){
-
-    return(
-      <View style={{paddingBottom:5}}>
-
-      {/* Separator */}
-        <View style={styles.separator} lightColor="#ffff" darkColor="rgba(255,255,255,0.1)" />
-  
-      {/* Button */}
-        <View style={styles.OutfitButton}> 
-          <Pressable  
-            // onPress={() => setModalVisible(true)}
-            >
-        <View style={{flex: 1,
-            flexDirection: 'row', 
-            alignItems:'center', 
-            justifyContent:'space-between',
-            backgroundColor: '#F0F0F0' //'#F194FF'
-            
-            }}>
-
-                {/* drag icon */}
-                <Icon
-                    name="drag"
-                    size={30}
-                    color="#C100E0"
-                    onPress={() => console.log('pressed')
-                }
-                />
-                {/* two lines drag icon */}
-                {/* <MaterialIcons 
-                    name="drag-handle" 
-                    size={24} 
-                    color="white" 
-                    style={styles.icon} 
-                    onPress={() => console.log('pressed')}
-                /> */}
-
-                <Text style={styles.OutfitButtonTextStyle}>Outfit</Text>
-
-                {/* trashcan icon */}
-                <Icon
-                    name="delete"
-                    size={24}
-                    color="#C100E0"
-                    onPress={() => console.log('pressed')}
-                />
-
-            </View>                
-
-          </Pressable>
-
-
-        </View>
-  
-  
-      </View>
-  
-    );
-  }
-
-
-function searchPlaceHolder(){
-const [number, onChangeNumber] = React.useState('');
-
-return(
-    <SafeAreaView>
-    <TextInput
-    style={{
-        paddingHorizontal: 5,
-        color:"black", 
-        // fontWeight:'bold'
-    }}
-        onChangeText={onChangeNumber}
-        value={number}
-        placeholder="Search"
-        // keyboardType="numeric"
-        placeholderTextColor="#9f9f9f"
-    />
-    </SafeAreaView>
-);
-}
   
 export default function EditOutfits() {
 
-    const [inputValue, setInputValue] = useState('');
     const navigation = useNavigation();
+
     const deleteOutfitButton = () => {
 
     }
@@ -124,259 +32,75 @@ export default function EditOutfits() {
         
     }
 
-    const handleCancel = () => {
-        setInputValue('');
-        // Navigate back to the previous screen or any desired screen
-        navigation.goBack(); // Assuming you're using a Stack Navigator
-    };
-    const handlePress = () => {
-        // Handle press event
-      };
-
-
-    // Update the title dynamically
-    React.useLayoutEffect(() => {
+    // Update the title dynamically to "Edit"
+    useLayoutEffect(() => {
         navigation.setOptions({
             title: 'Edit',
             headerTitleAlign: 'center',
 
         });
     }, [navigation]);
-
-
-
-
-
-    const [selectedType, setSelectedType] = useState(Clothe.Type.Unknown);
-    const [selectedColor, setSelectedColor] = useState(Clothe.Color.Unknown); // Default value
-    const [selectedSize, setSelectedSize] = useState(Clothe.SleeveSize.Unknown); // Default value
-    
-    const typeKeys = Object.keys(Clothe.Type).filter(key => isNaN(Number(key)));
-    const colorKeys = Object.keys(Clothe.Color).filter(key => isNaN(Number(key)));
-    const sleeveSizeKeys = Object.keys(Clothe.SleeveSize).filter(key => isNaN(Number(key)));
   
-
     return (
 
         <View style={styles.bigcontainer}> 
         {/* bigger container */}
         
-
         <View style={styles.fixedContainer}> 
         {/* Search & Filter container is inside this Fixed container */}
 
             {/* Search & Filter box*/}
             <View style={styles.searchAndFilter}>
 
-            {/* Search Bar */}
-            <View style={styles.searchContainer}>
-                {/* <Text style={styles.searchText}>Search</Text> */}
-                {searchPlaceHolder()}
-
-            </View>
-                <View style={styles.filterTextContainer}>              
-                    <Text style={styles.filterText}>Type</Text>
-                    <Text style={styles.filterText}>Color</Text>
-                    <Text style={styles.filterText}>Sleeve Size</Text>
-
+                {/* Search Bar */}
+                <View style={styles.searchContainer}>
+                    {searchPlaceHolder()}
                 </View>
-            {/* Filter boxes */}
-            <View style={styles.pickerBox}>
 
-                {/* <View style={styles.filterContainer}>   */}
-
-                        <Picker style={styles.picker} itemStyle={styles.pick}
-                            selectedValue={selectedType} 
-                            onValueChange={(itemValue: Clothe.Type) => 
-                                setSelectedType(itemValue) // Update the selectedType state 
-                        }>
-                        {typeKeys.map((typeKey) => (
-                            <Picker.Item
-                                key={typeKey}
-                                label={typeKey}
-                                value={Clothe.Type[typeKey as keyof typeof Clothe.Type]}
-                            />
-                            ))}
-                        </Picker>
-                {/* </View> */}
-
-
-
-                {/* <View style={styles.filterContainer}>   */}
-                    {/* <Text style={styles.filterText}>Color</Text> */}
-                        <Picker style={styles.picker} itemStyle={styles.pick}
-                            selectedValue={selectedColor} 
-                            onValueChange={(itemValue) => 
-                                setSelectedColor(itemValue)
-                            }>
-                            {colorKeys.map((colorKey) => (
-                                <Picker.Item
-                                    key={colorKey}
-                                    label={colorKey}
-                                    value={Clothe.Color[colorKey as keyof typeof Clothe.Color]}
-                                />
-                            ))}
-                        </Picker>
-                {/* </View> */}
-
-
-
-                {/* <View style={styles.filterContainer}>   */}
-                    {/* <Text style={styles.filterText}>Sleeve Size</Text> */}
-                        <Picker style={styles.picker} itemStyle={styles.pick}
-                            selectedValue={selectedSize}
-                            onValueChange={(itemValue, itemIndex) => 
-                                setSelectedSize(itemValue)
-                            }>
-                            {sleeveSizeKeys.map((sleeveSizeKey) => (
-                                <Picker.Item
-                                    key={sleeveSizeKey}
-                                    label={sleeveSizeKey}
-                                    value={Clothe.SleeveSize[sleeveSizeKey as keyof typeof Clothe.SleeveSize]}
-                                />
-                            ))}
-                        </Picker>
-                    {/* </View> */}
-
-                </View> 
-                {/* filterbox end*/}
+                {/* Filter Titles and Boxes */}
+                {outfitFilterBoxes()}
             
             </View> 
             {/* searchAndFilter end*/}
 
 
+        </View> 
+        {/* Fixed Container view */}
+
+        {/* smaller container with outfits buttons and save button */}
+        <View style={styles.smallercontainer}> 
+            <ScrollView>
+                <View style={styles.outfitsButtonContainer}> 
+
+                    {/* Outfit Buttons */}
+                    {addEditOutfitButton()}
+                    {addEditOutfitButton()}
+                    {addEditOutfitButton()}
+                    {addEditOutfitButton()}
+                    {addEditOutfitButton()}       
+                    {addEditOutfitButton()}
+                    {addEditOutfitButton()}  
+
+                </View> 
+                {/* Container with Outfits Buttons*/}
+            </ScrollView>
+
+
+            {/* Save & Cancel Button Container */}
+            <View style={styles.buttonContainer}>
+            {/* <View style={{alignItems:'center'}}> */}
+                {CancelButton()}
+                {SaveButton()}
+
+
             </View> 
-            {/* Fixed Container view */}
-
-
-
-
-
-            <View style={styles.smallercontainer}> 
-            {/* smaller container with both filter box + outfits buttons and save button */}
-
-        <ScrollView>
-
-        <View style={styles.outfitsButtonContainer}> 
-        {/* container with outfits buttons & save button */}
-
-
-        {/* First Button without separator */}
-            <View style={{paddingBottom:5}}>
-
-        {/* Outfit Buttons */}
-        
-        {/* First Button */}
-            <View style={styles.OutfitButton}>
-          <Pressable  
-            // onPress={() => setModalVisible(true)}
-            >
-        <View style={{flex: 1, 
-            flexDirection: 'row', 
-            alignItems: 'center', //'flex-start', 
-            justifyContent:'space-between',
-            backgroundColor: '#F0F0F0', 
-            // marginHorizontal:10
-            // '#F194FF'
-            }}>
-
-                {/* drag icon */}
-                <Icon
-                    name="drag"
-                    size={30}
-                    color="#C100E0"
-                    onPress={() => console.log('pressed')}
-                />
-                {/* two lines drag icon */}
-                {/* <MaterialIcons 
-                    name="drag-handle" 
-                    size={24} 
-                    color="white" 
-                    style={styles.icon} 
-                    onPress={() => console.log('pressed')}
-                /> */}
-
-                <Text style={styles.OutfitButtonTextStyle}>Outfit</Text>
-
-                {/* trashcan icon */}
-                <Icon
-                    name="delete"
-                    size={24}
-                    color="#C100E0"
-                    onPress={() => console.log('pressed')}
-                />
-
-            </View>                
-
-          </Pressable>
-          </View>                
-
-
-        </View>
-
-
-        {/* Button */}
-        {addOutfitButton()}
-        
-        {/* Button */}
-        {addOutfitButton()}
-
-        {/* Button */}
-        {addOutfitButton()}
-
-        {/* Button */}
-        {addOutfitButton()}
-
-        {/* Button */}
-        {addOutfitButton()}       
- 
-
-
-
-
-        {/* Save & Cancel Button */}
-        <View style={styles.buttonContainer}>
-
-
-            {/* cancel Button */}
-                {/* <View style={styles.cancelButton}> 
-                    <Pressable onPress={handleCancel} >
-                        <Text style={styles.cancelButtonText}>Cancel</Text>
-                    </Pressable>
-                </View> */}
-
-
-            {/* Separator */}
-            <View style={styles.separator2} lightColor="#ffff" darkColor="rgba(255,255,255,0.1)" />
-
-
-            {/* save Button */}
-                <View style={styles.saveButton}> 
-                    <Pressable onPress={handleCancel} >
-                        <Text style={styles.saveButtonText}>Save</Text>
-                    </Pressable>
-                </View>
-
+            {/* Save Button Container */} 
 
         </View> 
-        {/* Save Button Container */}        
-        
+        {/* Container with Outfit Buttons and Save Button*/}
 
         </View> 
-        {/* smaller container with Outfits Buttons & Save Button*/}
-        
-                </ScrollView>
-
-
-        </View> 
-        {/* smaller container with Filter Box + Outfit Buttons and Save Button*/}
-
-
-
-        </View> 
-        //{/* bigger container */}
-        
-        
+        //{/* bigger container */}   
     );
 }
 
@@ -384,37 +108,28 @@ const styles = StyleSheet.create({
     bigcontainer: { //all components are in this parent container
         flex: 1,
         // backgroundColor: 'black', //ffff
-
-        // alignItems: 'center',
-        // justifyContent: 'center',
-        // flexDirection: 'column',
-        // flexWrap: 'wrap',
-        // width: '100%',
-        // marginVertical: 20,
         },
         
     smallercontainer: { //all components are inside this smaller container
         flex: 1,
         // backgroundColor: 'pink',
         padding: 10, 
-        margin: 5, //margin of the square - how big is it
-        borderRadius: 20, //rounds edges
-        top:100,
+        margin: 5,
+        borderRadius: 20,
+        top:110,
         marginBottom:110,
         zIndex: 1, // Ensure it appears above other content
         ...Platform.select({
           ios: {
               width: "95%",
-              top:120
+              top:150
           },
           android: {
               width: "95%",
-              top:120,
+              top:150,
           },
       }),
     },
-
-
 
     // Search components styles
     fixedContainer: {
@@ -422,16 +137,14 @@ const styles = StyleSheet.create({
         top: 0,
         left: 0,
         right: 0,
-        // bottom:10, //Nope, covers the whole page
-        // backgroundColor: 'lightblue',
-        // zIndex: 1, // Ensure it appears above other content
+
         ...Platform.select({
           ios: {
-              height: "20%",
+              height: "35%",
               width: "95%",
           },
           android: {
-              height: "27%",
+              height: "35%",
               width: "100%",
               // top:5
           },
@@ -442,9 +155,8 @@ const styles = StyleSheet.create({
         // flex: 1,
         backgroundColor: '#D9D9D9',
         padding: 5, 
-        margin: 10, //margin of the square - how big is it
-        borderRadius: 10, //rounds edges
-        // marginTop: 1,
+        margin: 10,
+        borderRadius: 10,
         marginVertical: 1,
     },
 
@@ -454,14 +166,12 @@ const styles = StyleSheet.create({
 
     searchAndFilter:{  //search and filters are inside this container
         backgroundColor: "#F0F0F0",
-        // alignItems: 'center',
         alignContent: 'space-between',
         justifyContent: 'space-evenly',
         borderRadius: 20,
         margin:10,
         padding: 10,
-        // marginTop:1,
-        elevation: 2, 
+
         shadowColor: '#000',
         shadowOffset: {
           width: 0.9,
@@ -469,6 +179,8 @@ const styles = StyleSheet.create({
         },
         shadowOpacity: 0.25,
         shadowRadius: 4, 
+        elevation: 2, 
+
         ...Platform.select({
             ios: {
                 height: "20%",
@@ -481,130 +193,15 @@ const styles = StyleSheet.create({
             },
         }),
     },
-
-    filterContainer:{  //filterbox and text label are inside this container
-    alignItems: 'center',
-    flexDirection: 'column',
-    backgroundColor: '#F0F0F0'
-    
-    },
-
-    filterTextContainer:{
-        flexDirection:'row', 
-        justifyContent:'space-evenly', 
-        backgroundColor: "#F0F0F0", 
-        left:5,
-        ...Platform.select({
-          ios: {
-              width: "33%",
-              bottom: 55,
-          },
-          android: {
-              width: "100%",
-              height: "25%",
-              top: 7,
-              left: 0,
-          },
-        }),
-      
-      },
-
-    filterText: { //styling for *text* on top of filter dropdowns
-    // paddingTop: 10,
-    paddingHorizontal: 10,
-    color:"black", 
-    fontWeight:'bold'
-    },
-
-    pickerBox: { //container where pickers are located
-        flexDirection: "row",
-        justifyContent: 'space-evenly',
-        // flex:1,
-        height: "45%",
-        backgroundColor:  'rgba(52, 52, 52, 0)', //'lightblue',
-        ...Platform.select({
-          ios: {
-              width: "33%",
-              bottom: 55,
-          },
-          android: {
-              width: "100%",
-              height: "55%",
-              bottom: 0,
-          },
-      }),     
-    },
-    picker: { // picker outside design
-        borderColor: 'black',
-        borderRadius: 10,
-
-        ...Platform.select({
-            ios: {
-                width: "33%",
-                bottom: 55,
-            },
-            android: {
-                width: "40%",
-                height: "55%",
-                bottom: 0,
-            },
-        }),
-    },
-    pick: { // picker design once an item is picked??? Honestly no idea
-        fontSize: 15,
-        ...Platform.select({
-            ios: {fontSize: 15},
-            android: {fontSize: 0,},
-        }),
-        
-    },
-
-
-
-
-
-
-
-
-
-
-
-
-
-    draggableContainer: {
-        width: '80%',
-        height: 200,
-        backgroundColor: '#ccc',
-      },
-      draggable: {
-        width: 50,
-        height: 50,
-        backgroundColor: 'blue',
-        borderRadius: 25,
-        justifyContent: 'center',
-        alignItems: 'center',
-      },
-      draggableInner: {
-        width: 40,
-        height: 40,
-        backgroundColor: 'white',
-        borderRadius: 20,
-      },
-
-
-
-    // Outfit Buttons component styles
+    // Outfit Buttons Container
     outfitsButtonContainer:{ // all outfits buttons AND save button are inside this container
         flex: 1,
         // backgroundColor: 'lightblue',
         padding: 10, 
-        margin: 5, //margin of the square - how big is it
-        borderRadius: 20, //rounds edges
-        // marginBottom:120,
-        // top:100,
-        // zIndex: 1, // Ensure it appears above other content
-        // maxHeight:480,
-        // paddingBottom: 100,
+        margin: 5, 
+        borderRadius: 20, 
+        marginTop:-20,
+
         ...Platform.select({
           ios: {
               height: "50%",
@@ -617,81 +214,16 @@ const styles = StyleSheet.create({
           },
       }),
     },
-
-
-    separator: { //added this to separate each button a smidge, is added just before a new button
-        marginVertical: 10,
-        height: .5,
-        width: '80%',
-      },
-    separator2: { //added this to create a space between cancel and save buttons
-        marginVertical: 20,
-        height: 10,
-        width: '3%',
-    },
-
-    OutfitButton: {
-        borderRadius: 15,
-        padding: 17,
-        elevation: 2,
-        backgroundColor: '#F0F0F0', //'#F194FF', //'#f8f4f4'
-        shadowColor: '#000',
-        shadowOffset: {
-          width: 0.9,
-          height: 2,
-        },
-        shadowOpacity: 0.25,
-        shadowRadius: 4,   
-        // maxHeight:60, 
-      },
-
-    OutfitButtonTextStyle: {
-        color: '#C100E0', //'white', //'#c404f0' //'#F194FF'
-        fontWeight: 'bold',
-        textAlign: 'center',
-        // justifyContent: 'center',
-      },
-
     // Cancel & Save Buttons
     buttonContainer:{
         flex: 1,
-        // alignItems: 'center',
-        // justifyContent: 'center',
+        justifyContent: 'space-around',
         flexDirection: 'row',
-        // flexWrap: 'wrap',
-        // backgroundColor: 'black',
-        margin: 10, //margin of the square - how big is it
-        justifyContent:'flex-end',
-        marginHorizontal:10,
-    
-    },
-    cancelButton: {
-        borderRadius: 10,
-        padding: 10,
-        height:'40%',
-        backgroundColor: 'red',
-        justifyContent: 'center'
+        backgroundColor: "#F0F0F0", 
+        margin: 10, 
 
+        position: 'absolute',
+        bottom:50,
+        left: 110
     },
-    cancelButtonText: {
-        color: 'white',
-        textAlign: 'center',
-        fontWeight: 'bold'
-    },
-
-    saveButton: {
-        borderRadius: 10,
-        padding: 10,
-        // height:'40%',
-        maxHeight:35,
-        backgroundColor: 'limegreen',
-        justifyContent: 'center'
-    },
-    saveButtonText: {
-        color: 'white',
-        textAlign: 'center',
-        fontWeight: 'bold',
-        
-    },
-
 });
